requirements.txt
certifi==2016.2.28
click==6.7
Flask==0.12.2
gunicorn==19.1.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.1
olefile==0.44
PasteDeploy==1.5.2
Pillow==4.2.1
requests==2.14.2
Werkzeug==0.12.2
